services/users/requirements.txt
pymongo
strawberry-graphql
fastapi
uvicorn
bcrypt
pyjwt
